implement Menu MVP style
