+++ teamgreen-pms-app/src/main/java/ch/bfh/bti7081/s2017/green/ui/components/menu/Menu.java
@@ -1,0 +1,7 @@
+package ch.bfh.bti7081.s2017.green.ui.components.menu;
+
+/**
+ * Created by Lukas on 15.06.2017.
+ */
+public class Menu {
+}
