+++ teamgreen-pms-app/src/main/java/ch/bfh/bti7081/s2017/green/ui/components/menu/MenuView.java
@@ -1,0 +1,10 @@
+package ch.bfh.bti7081.s2017.green.ui.components.menu;
+
+import green.auth.UserContext;
+
+/**
+ * Created by Lukas on 15.06.2017.
+ */
+public interface MenuView {
+    void updateUserMenu (UserContext userContext);
+}
