+++ teamgreen-pms-app/src/main/webapp/VAADIN/themes/dashboard/addons.scss
